+++ ode/Development/webapp/src/main/java/com/bah/ode/model/OdeTstRequest.java
@@ -1,5 +1,7 @@
 package com.bah.ode.model;
 
+import java.math.BigDecimal;
+
 public class OdeTstRequest extends OdeRequest {
    private static final long serialVersionUID = -5219513940183321970L;
 
@@ -7,8 +9,8 @@ public class OdeTstRequest extends OdeRequest {
       super();
    }
 
-   public OdeTstRequest(Double nwLat, Double nwLon, Double seLat, 
-         Double seLon, OdeDataType dataType) {
+   public OdeTstRequest(BigDecimal nwLat, BigDecimal nwLon, BigDecimal seLat, 
+         BigDecimal seLon, OdeDataType dataType) {
       super(nwLat, nwLon, seLat, seLon, OdeRequestType.Test, dataType);
    }
